+++ api/src/main/java/org/searchisko/api/security/jaas/ProviderLoginModule.java
@@ -33,9 +33,8 @@ import org.searchisko.api.util.CdiHelper;
  */
 public class ProviderLoginModule extends UsernamePasswordLoginModule {
 
-	@Inject
-	protected Logger log;
-
+	protected Logger log = Logger.getLogger(getClass().getName());
+	
 	@Inject
 	protected ProviderService providerService;
 
@@ -43,6 +42,14 @@ public class ProviderLoginModule extends UsernamePasswordLoginModule {
 	public void initialize(Subject subject, CallbackHandler callbackHandler, Map<String, ?> sharedState, Map<String, ?> options) {
 		try {
 			CdiHelper.programmaticInjection(ProviderLoginModule.class, this);
+			
+			//	Since login modules are not managed beans it's not possible to use annotations like @Inject or @Resource by default.
+			//	Therefore the above approach handles this in a creative way. Just in case it stops working in the future
+			//	those lines below can replicate its functionality in a traditional, JNDI lookup-based way.
+			//	InitialContext initialContext = new InitialContext();
+			//	Object lookup = initialContext.lookup("java:module/ProviderService");
+			//	this.providerService = (ProviderService) lookup;
+			
 		} catch (NamingException e) {
 			throw new RuntimeException("Cannot initialize Login module", e);
 		}